print crypto lib versions and stuff

+++ crypto.h
@@ -3,7 +3,7 @@
 
 #include "common.h"
 
-void sha1(u8 *out, const u8 *in, size_t len);
+void sha1(u8 *out, const u8 *in, unsigned len);
 
 void aes_128_ecb_init();
 void aes_128_ecb_set_key(const u8 *key);
